+++ src/gnunet/service/revocation/pow_test.go
@@ -117,7 +117,7 @@ func TestRevocationRFC(t *testing.T) {
 	sigBlock := &SignedRevData{
 		Purpose: &crypto.SignaturePurpose{
 			Size:    uint32(20 + revData.ZoneKeySig.KeySize()),
-			Purpose: uint32(enums.SIG_REVOCATION),
+			Purpose: enums.SIG_REVOCATION,
 		},
 		Timestamp: revData.Timestamp,
 		ZoneKey:   &revData.ZoneKeySig.ZoneKey,
